+++ src/sanityClient.js
@@ -4,7 +4,9 @@ const client = createClient({
 	projectId: 'tun1nqnk',
 	dataset: 'production',
 	apiVersion: '2022-08-21',
-	useCdn: true
+	useCdn: true,
+	perspective: 'published',
+	stega: false
 });
 
 export default client;